% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0asub_covariantMatrix.R
\name{covariantMatrix}
\alias{covariantMatrix}
\title{Covariant Matrix}
\usage{
covariantMatrix(meta)
}
\arguments{
\item{meta:}{list of many attributes}
}
\value{
list of 2 elements: matrix of covariants and means (not sure)
}
\description{
function to generate covariant Matrix
}
\examples{

}
\seealso{
Other functions used to read covariates: 
\code{\link{extractWhoData}()},
\code{\link{getCovariateInfo}()},
\code{\link{v2015ct}()}
}
\concept{functions used to read covariates}
