% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_for_each_observation.R
\name{dm_for_each_observation}
\alias{dm_for_each_observation}
\title{dm_for_each_observation}
\usage{
dm_for_each_observation(main_data, dm_rules = dm_rules())
}
\arguments{
\item{dm_rules}{}
}
\value{
dm_name_i
}
\description{
find the data model for each observation. requires the main data to be a list
}
