% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documents.R
\name{document_data}
\alias{document_data}
\title{document_data}
\usage{
document_data(df)
}
\arguments{
\item{df}{}
}
\description{
automatically generates roxygen skeleton for data use base r's "str" function
}
