% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0asub_extractWhoData.R
\name{extractWhoData}
\alias{extractWhoData}
\title{Extract Who Data}
\usage{
extractWhoData(lifetable_path, meta)
}
\arguments{
\item{lifetable_path:}{lifetable}

\item{meta:}{list of many attributes}
}
\value{
m: list of 4 attributes
}
\description{
function to extract Who Data
}
\examples{

}
\seealso{
Other functions used to read covariates: 
\code{\link{covariantMatrix}()},
\code{\link{getCovariateInfo}()},
\code{\link{v2015ct}()}
}
\concept{functions used to read covariates}
