% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_global_meta.R
\name{filter_global_meta}
\alias{filter_global_meta}
\title{obtain_country_meta_info}
\usage{
filter_global_meta(iso_alpha_3_code, meta)
}
\arguments{
\item{runname}{}
}
\value{

}
\description{
obtain_country_meta_info
}
