% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bmat_onecountry.R
\name{fit_bmat_onecountry}
\alias{fit_bmat_onecountry}
\title{fit_bmat_onecountry}
\usage{
fit_bmat_onecountry(
  tar_depends = NULL,
  runname = NULL,
  main_data,
  meta = bmat::meta,
  diagnostic = FALSE,
  constraint = FALSE,
  quantiles = c(0.1, 0.5, 0.9),
  getarrs = TRUE,
  pars_to_save = c("mu.ct", "oneminpi.c", "logphi.jnew", "tau.jnew")
)
}
\description{
fit_bmat_onecountry
}
