% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0asub_getCovariateInfo.R
\name{getCovariateInfo}
\alias{getCovariateInfo}
\title{Get Covariate Info}
\usage{
getCovariateInfo(gdpfile, wppfile, sabfile, meta)
}
\arguments{
\item{gdpfile.}{GDP}

\item{sabfile:}{2015 SAB database}

\item{wpp2015est:}{WPP2015 estimates}

\item{meta:}{list of many attributes}
}
\value{
m: list of 4 attributes
}
\description{
function to get covariate Info
}
\examples{

}
\seealso{
Other functions used to read covariates: 
\code{\link{covariantMatrix}()},
\code{\link{extractWhoData}()},
\code{\link{v2015ct}()}
}
\concept{functions used to read covariates}
