% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documents.R
\docType{data}
\name{main_data}
\alias{main_data}
\title{main_data}
\format{
data.frame':	3675obs.of33variables
\describe{
\item{iso}{\emph{\sQuote{Character}} ...... }
\item{year}{\emph{\sQuote{Numeric}} ...... }
\item{start}{\emph{\sQuote{Numeric}} ...... }
\item{end}{\emph{\sQuote{Numeric}} ...... }
\item{obs_matdeaths}{\emph{\sQuote{Numeric}} ...... }
\item{final_pm}{\emph{\sQuote{Numeric}} ...... }
\item{final_env}{\emph{\sQuote{Numeric}} ...... }
\item{shortcite}{\emph{\sQuote{Character}} ...... }
\item{modelinclude}{\emph{\sQuote{Logical}} ...... }
\item{modelinclude_reason}{\emph{\sQuote{Character}} ...... }
\item{ss_toedit}{\emph{\sQuote{Character}} ...... }
\item{completeness_inq}{\emph{\sQuote{Numeric}} ...... }
\item{obs_selogpm}{\emph{\sQuote{Numeric}} ...... }
\item{type}{\emph{\sQuote{Character}} ...... }
\item{definition}{\emph{\sQuote{Character}} ...... }
\item{rhovrfinal_bmat}{\emph{\sQuote{Numeric}} ...... }
\item{usa}{\emph{\sQuote{Numeric}} ...... }
\item{obs_env}{\emph{\sQuote{Numeric}} ...... }
\item{sens}{\emph{\sQuote{Numeric}} ...... }
\item{var_sens}{\emph{\sQuote{Numeric}} ...... }
\item{sens_sq}{\emph{\sQuote{Numeric}} ...... }
\item{oneminsens_sq}{\emph{\sQuote{Numeric}} ...... }
\item{spec}{\emph{\sQuote{Numeric}} ...... }
\item{var_spec}{\emph{\sQuote{Numeric}} ...... }
\item{spec_sq}{\emph{\sQuote{Numeric}} ...... }
\item{oneminspec_sq}{\emph{\sQuote{Numeric}} ...... }
\item{rho_sesp}{\emph{\sQuote{Numeric}} ...... }
\item{cov_sesp}{\emph{\sQuote{Numeric}} ...... }
\item{hasstudies}{\emph{\sQuote{Logical}} ...... }
\item{mmr_observed}{\emph{\sQuote{Numeric}} ...... }
\item{final_mmr}{\emph{\sQuote{Numeric}} ...... }
\item{final_pm_before_crisis}{\emph{\sQuote{Numeric}} ...... }
\item{final_mmr_before_crisis}{\emph{\sQuote{Numeric}} ...... }
}
}
\source{

}
\usage{
main_data
}
\description{
contains data that is central to BMat
}
\references{

}
\keyword{datasets}
