% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documents.R
\docType{data}
\name{mics_mmr}
\alias{mics_mmr}
\title{mics_mmr}
\format{
A data frame with 49 variables:
\describe{
  \item{New data as of June/July 2019=1}{\emph{\sQuote{Numeric}} explain the variable here. }
  \item{AB 6 March}{\emph{\sQuote{Character}} ...... }
  \item{who 5 march 2019}{\emph{\sQuote{Character}} ....... }
  \item{response la}{\emph{\sQuote{Character}} .......... }
  \item{5 march response2 to LA }{\emph{\sQuote{Character}} ................}
  \item{AB comments}{\emph{\sQuote{Character}}................ }
  \item{Data before 2000 or mid-year not 2000: 1= yes}{\emph{\sQuote{Numeric}} ............   }
  \item{15-19 years; yes=1, no=2 (>2000)}{\emph{\sQuote{Character}} ...... }
}
}
\source{

}
\usage{
mics_mmr
}
\description{
A dataset containing mmr data from MICS surveys
}
\references{

}
\keyword{datasets}
