% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bmat_wrappers.R
\name{plot_bmat_global}
\alias{plot_bmat_global}
\title{plot_bmat_global}
\usage{
plot_bmat_global(runname = NULL, estimates_to_compare, iso_alpha_3_codes)
}
\arguments{
\item{iso_alpha_3_codes}{}
}
\value{

}
\description{
plot_bmat_global
}
