% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bmat_wrappers.R
\name{plot_bmat_onecountry_sub}
\alias{plot_bmat_onecountry_sub}
\title{plot_bmat_onecountry_sub}
\usage{
plot_bmat_onecountry_sub(
  main_data_adjusted,
  estimates_onecountry,
  estimates_to_compare,
  iso_alpha_3_code
)
}
\arguments{
\item{fit_list}{}
}
\value{

}
\description{
plot_bmat_onecountry_sub
}
