% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a_process_precrisis_meta.R
\name{process_precrisis_meta}
\alias{process_precrisis_meta}
\title{process_meta}
\usage{
process_precrisis_meta(lifetable_path, gdp_path, sab_path, wpp_path)
}
\arguments{
\item{lifetable_path}{}

\item{gdp_path}{}

\item{sab_path}{}

\item{wpp_path}{}

\item{info_path}{}

\item{updated_names_path}{}
}
\description{
function to process covariates and other descriptive data
}
