% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_se_sp_tibble 2.R
\name{summarize_se_sp_tibble}
\alias{summarize_se_sp_tibble}
\title{summarize_se_sp_tibble}
\usage{
summarize_se_sp_tibble(se_samples.st, sp_samples.st, maxvar_se, maxvar_sp)
}
\arguments{
\item{maxvar_sp}{}
}
\value{
output: a tibble with median/var/corr and other measures reported for both se and sp
}
\description{
Takes in samples of se and sp for all years, summarizes into medians and vars for each, calculates correlation between se and sp by year, outputs a tibble
with combined median, var and correlation estimates for both se and sp
}
