% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0asub_v2015ct.R
\name{v2015ct}
\alias{v2015ct}
\title{v2015ct}
\usage{
v2015ct(meta)
}
\arguments{
\item{meta:}{list of many attributes}
}
\value{
v2015.ct: matrix of dimensions (meta$C,meta$nyears)
}
\description{
function to generate v2015.ct
}
\examples{

}
\seealso{
Other functions used to read covariates: 
\code{\link{covariantMatrix}()},
\code{\link{extractWhoData}()},
\code{\link{getCovariateInfo}()}
}
\concept{functions used to read covariates}
